fix(android): русский интерфейс, аватарки и загрузка чатов с голосовыми
Три проблемы мобильного клиента:

1. Полностью русский UI независимо от языка системы: resourceConfigurations
   += "ru" — английские ресурсы values-en/ исключаются из сборки, всегда
   используется русский дефолт values/.

2. Аватарки других пользователей не грузились: бэкенд отдаёт avatarUrl
   относительным (/media/avatars/…), а Coil получал сырой путь. Прогоняем
   через MediaUrlBuilder.absoluteOrNull в списке чатов, шапке переписки и
   оверлее звонка. Заодно шапка переписки теперь показывает аватар+имя
   собеседника (как в вебе).

3. «Не удалось загрузить сообщения» в части чатов: бэкенд отдаёт waveform
   голосовых/кружков как нормализованные Float 0..1, а DTO/модели клиента
   ждали List<Int> — kotlinx.serialization падал на дробных значениях, и весь
   чат с голосовым не грузился. Перевели waveform на List<Float> везде
   (DTO сообщений/загрузки, доменная модель, Room-JSON, рендер волны).
   Плюс тап по ошибке истории теперь повторяет загрузку.

+++ mobile/android/app/src/main/java/com/infy/messenger/feature/chat/data/ChatMappers.kt
@@ -44,7 +44,7 @@ private data class AttachmentJson(
     val width: Int?,
     val height: Int?,
     val durationMs: Int?,
-    val waveform: List<Int>,
+    val waveform: List<Float>,
     val transcript: String?,
     val listenedAt: Long?,
 )
